refactor(codegen): extract fuzz routing and paren policy modules
Phase 1 Batch 4 (Moves 10-11 from docs/PHASE1-PLAN.md, the final
Phase 1A batch).  Extract the fuzz-routing helpers and the
operator-precedence / parenthesisation policy from latex_gen.py
into two new mixin modules.

Moves are byte-identical Fowler Move-Method / Extract-Class
refactorings.

* fuzz_routing.py — _DAT_EXPRESSION_TYPES (ClassVar) and
  _expression_contains_dat_construct.  These decide whether an
  abbreviation right-hand side belongs inside a fuzz Z-environment
  or in noindent inline math.
* paren_policy.py — the four operator-precedence class-vars
  (BINARY_OPS, PRECEDENCE, RIGHT_ASSOCIATIVE, UNARY_PRECEDENCE),
  the binary-operator text mapper (_map_binary_operator), and the
  three parens-needed predicates (_needs_parens,
  _body_has_connective, _quantifier_needs_parens).  This module is
  the foundation Phase 2 will build on.

The CodegenDispatch TYPE_CHECKING block grows by one entry
(_in_argue_block) so the moved _needs_parens type-checks.

Gate 4 verified:
* make check — clean.
* make test-e2e — 159/159 byte-identical.
* make refactor-diff — 190/190 byte-identical vs main.

latex_gen.py shrinks from 2871 → 2514 lines.  Cumulative reduction
from the original 6864-line monolith: 63.4%.  Phase 1A complete.
