+++ src/txt2tex/codegen/__init__.py
@@ -7,8 +7,10 @@ from txt2tex.codegen import (
     algebra,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     bindings,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     expressions,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    fuzz_routing,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     overflow,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     paragraphs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    paren_policy,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     proofs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     schemas,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     text_blocks,  # noqa: F401  # pyright: ignore[reportUnusedImport]
